+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+- `multi_setup_poser` tests
+
+
 ## [0.5.0] - 2024-04-09
 
 ### Added